move annual report and contact pages
