+++ annual_reports/models.py
@@ -1,0 +1,115 @@
+from core.models import (
+    BasicPageAbstract,
+    FeatureablePageAbstract,
+    SearchablePageAbstract,
+)
+from django.core.validators import MaxValueValidator, MinValueValidator
+from django.db import models
+from wagtail.admin.edit_handlers import (
+    FieldPanel,
+    MultiFieldPanel,
+)
+from wagtail.core.models import Page
+from wagtail.documents.edit_handlers import DocumentChooserPanel
+from wagtail.images.edit_handlers import ImageChooserPanel
+
+
+class AnnualReportListPage(BasicPageAbstract, Page):
+    max_count = 1
+    parent_page_types = ['core.BasicPage']
+    subpage_types = ['annual_reports.AnnualReportPage']
+    templates = 'annual_reports/annual_report_list_page.html'
+
+    content_panels = [
+        BasicPageAbstract.title_panel,
+        BasicPageAbstract.body_panel,
+        BasicPageAbstract.images_panel,
+    ]
+    settings_panels = Page.settings_panels + [
+        BasicPageAbstract.submenu_panel,
+    ]
+
+    class Meta:
+        verbose_name = 'Annual Report List Page'
+
+
+class AnnualReportPage(FeatureablePageAbstract, Page, SearchablePageAbstract):
+    """View annual report page"""
+
+    image_poster = models.ForeignKey(
+        'wagtailimages.Image',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+        verbose_name='Cover image',
+        help_text='Poster sized image that is displayed in the featured section on the Annual Reports page.',
+    )
+    report_english = models.ForeignKey(
+        'wagtaildocs.Document',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+    )
+    report_financial = models.ForeignKey(
+        'wagtaildocs.Document',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+    )
+    report_french = models.ForeignKey(
+        'wagtaildocs.Document',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+    )
+    report_interactive = models.CharField(
+        blank=True,
+        max_length=255,
+        help_text='Internal path to the interactive report. Example: /interactives/2019annualreport',
+    )
+    year = models.IntegerField(validators=[MinValueValidator(2005), MaxValueValidator(2050)])
+
+    # Reference field for the Drupal-Wagtail migrator. Can be removed after.
+    drupal_node_id = models.IntegerField(blank=True, null=True)
+
+    content_panels = Page.content_panels + [
+        MultiFieldPanel(
+            [
+                FieldPanel('year'),
+            ],
+            heading='General Information',
+            classname='collapsible',
+        ),
+        MultiFieldPanel(
+            [
+                DocumentChooserPanel('report_english'),
+                DocumentChooserPanel('report_french'),
+                DocumentChooserPanel('report_financial'),
+                FieldPanel('report_interactive'),
+            ],
+            heading='Reports',
+            classname='collapsible',
+        ),
+        MultiFieldPanel(
+            [
+                ImageChooserPanel('image_poster'),
+            ],
+            heading='Images',
+            classname='collapsible collapsed',
+        )
+    ]
+    promote_panels = Page.promote_panels + [
+        FeatureablePageAbstract.feature_panel,
+        SearchablePageAbstract.search_panel,
+    ]
+    parent_page_types = ['annual_reports.AnnualReportListPage']
+    subpage_types = []
+    templates = 'annual_reports/annual_report_page.html'
+
+    class Meta:
+        verbose_name = 'Annual Report Page'
+        verbose_name_plural = 'Annual Report Pages'
