+++ cigionline/settings/base.py
@@ -26,9 +26,11 @@ BASE_DIR = os.path.dirname(PROJECT_DIR)
 # Application definition
 
 INSTALLED_APPS = [
+    'annual_reports',
     'articles',
     'careers',
     'compressor',
+    'contact',
     'core',
     'events',
     'home',
