+++ home/models.py
@@ -90,8 +90,8 @@ class HomePage(Page):
         'articles.ArticleSeriesPage',
         'articles.MediaLandingPage',
         'careers.JobPostingListPage',
+        'contact.ContactPage',
         'core.BasicPage',
-        'core.ContactPage',
         'core.PrivacyNoticePage',
         'events.EventListPage',
         'multimedia.MultimediaListPage',
